Add Flex Asset Resources and New/Set/Remove Functionality
Added the following Flex Asset-related endpoints:
 - Flexible Assets
 - Flexible Asset Types
 - Flexible Asset Fields

Additionally, functionality (on all resources) was added for creating,
updating, and deleting (based on API support).

Creating and updating is based on the assumption that the hashtable
passed to the API is correctly formed. For example,
`Update-ITGlueUsers` has the required `data[type]` and
`data[attributes]` fields given in the `$data` variable passed to the
function. These data structures and required fields are described
in-depth by the documentation present at
https://api.itglue.com/developer

—

Feature parity should now be present with the current version of IT
Glue’s API.

—

More testing needs to be done before merging into Master.