Improve New/Set/Remove Functionality & Fix Errors
The functionality for passing and parsing data to the `New-`, `Set-`,
and `Remove-` functions has been improved. Now you can pass a variable
with hashes of whatever data is expected by each API endpoint and the
functions will automatically manage conversion to JSON.

Some errors/bugs were fixed that would cause a few functions to fail.

A new variable was added (`$ITGlue_JSON_Conversion_Depth`) to support
conversion to JSON by setting the level of depth for the
`ConvertTo-Json` cmdlet globally for the module. Currently, the default
value is `100`, which is more than enough to support the level of
nesting required for API usage.

—

Functions have been tested and are ready to be pulled merged into
Master.