Support full functionality (#1)
* Add Flex Asset Resources and New/Set/Remove Functionality

Added the following Flex Asset-related endpoints:
 - Flexible Assets
 - Flexible Asset Types
 - Flexible Asset Fields

Additionally, functionality (on all resources) was added for creating,
updating, and deleting (based on API support).

Creating and updating is based on the assumption that the hashtable
passed to the API is correctly formed. For example,
`Update-ITGlueUsers` has the required `data[type]` and
`data[attributes]` fields given in the `$data` variable passed to the
function. These data structures and required fields are described
in-depth by the documentation present at
https://api.itglue.com/developer

—

Feature parity should now be present with the current version of IT
Glue’s API.

—

The functionality for passing and parsing data to the `New-`, `Set-`,
and `Remove-` functions has been improved. Now you can pass a variable
with hashes of whatever data is expected by each API endpoint and the
functions will automatically manage conversion to JSON.

Some errors/bugs were fixed that would cause a few functions to fail.

A new variable was added (`$ITGlue_JSON_Conversion_Depth`) to support
conversion to JSON by setting the level of depth for the
`ConvertTo-Json` cmdlet globally for the module. Currently, the default
value is `100`, which is more than enough to support the level of
nesting required for API usage.

—

Functions have been tested and are ready to be pulled merged into
Master.