Add Support for Passwords API
With the release of API v1.8.0, a new Passwords endpoint was added.
This update uncovers that endpoint in the PowerShell module with the
addition of the following functions:
 - `New-ITGluePasswords`
 - `Get-ITGluePasswords`
 - `Set-ITGluePasswords`

Release notes for the Passwords api are available
[here](https://kb.itglue.com/hc/en-us/articles/360001797531-Release-Note
s-1-8-0-API).

——————

The Module Version was also updated to reflect the API version it
supports.