Add support for EU API endpoints
API endpoints have been added for deployments in the EU datacenter.
This update uncovers the ability to use the EU API links by use of a
`-data_center` (or `-dc`) option on the `Add-ITGlueBaseURI` function.
This optional argument takes `EU` and `US` as options, however the US
datacenter is still kept as the default even if no arguments are given.

—

HTTPS is also set by default for both the US and EU data centers.