v2.0.0 Release (#28)
* Update development branch with current proposed updates (#22)

* Add Support for Passwords API

With the release of API v1.8.0, a new Passwords endpoint was added.
This update uncovers that endpoint in the PowerShell module with the
addition of the following functions:
 - `New-ITGluePasswords`
 - `Get-ITGluePasswords`
 - `Set-ITGluePasswords`

Release notes for the Passwords api are available
[here](https://kb.itglue.com/hc/en-us/articles/360001797531-Release-Note
s-1-8-0-API).

——————

The Module Version was also updated to reflect the API version it
supports.

* Add support for EU API endpoints

API endpoints have been added for deployments in the EU datacenter.
This update uncovers the ability to use the EU API links by use of a
`-data_center` (or `-dc`) option on the `Add-ITGlueBaseURI` function.
This optional argument takes `EU` and `US` as options, however the US
datacenter is still kept as the default even if no arguments are given.

—

HTTPS is also set by default for both the US and EU data centers.

* Added enhanced filtering capability

Most endpoints (especially Index/Show, and Update) now allow for
expanded filtering capabilities. This updated exposes those new
capabilities as command parameters.

* Fixed Configurations Return Data

Fixed a bug that was pointed out that caused incorrect data to be
returned from the `Get-ITGlueConfigurations’ function.

* Compatability Update

All resources now match the current API spec given at
https://api.itglue.com/developer.

This includes enhanced bulk update, bulk destroy, and filtering
capabilities.

Thanks to @TheMattCollins0 for changes to make adding an API key via
command line easier.

* Added file to resolve Import-Module HRESULT: 0x8013104

* Support Nested Relationships Route for Show

* Update Documents in Development branch (#25)

* Markdown formatting edits and update doc for additional commands

* Markdown formatting edits

* Update to Current API Standards

This update overhauls all endpoints to bring them current with the IT Glue API. Any and all filtering, options, and endpoints are covered.

:warning: This is a **breaking** update.  In order to better match the API docs that IT Glue presents at https://api.itglue.com/development, the `$data` input taken for module functions will *only* consist of the information present in the `data` array, rather than requiring what was effectively a `$body` request with a data array nested as a value.

Some other styalistic choices were propogated throughout the code as well. For example, the `-important` parameter that some functions support now is a boolean. You no longer have to pass `0` or `1`, but can also pass `$true`, `$false`, or any other PowerShell boolean equivalent.

* Minor ITGlueAPI.psd1 file edits (#27)

* Markdown formatting edits and update doc for additional commands

* Markdown formatting edits

* Intial check in of basic tests for the module

* Expanded testing

* Minor edits

* Revert "Expanded testing"

This reverts commit 6c8f31eecbae9c2d17cfe303dcd96c7a3461e871.

* Revert "Intial check in of basic tests for the module"

This reverts commit ead9f258c5891d8f56ec1d6c7fe54f877e09e11a.

* Update to 2.0.0

The module version has been updated to `2.0.0`.

From now on, Semantic Versioning (https://semver.org) should be followed on the module.

Given a version number MAJOR.MINOR.PATCH, increment the:

MAJOR version when you make incompatible API changes,
MINOR version when you add functionality in a backwards-compatible manner, and
PATCH version when you make backwards-compatible bug fixes.

* Implement Error Handling on REST Headers

This implements the fix suggested by @ecspresso (thanks!!) and closes #29.

Co-Authored-By: Emile <ecspresso@users.noreply.github.com>